+++ src/app/model.rs
@@ -8,7 +8,10 @@ pub enum WarcMessage {
     Header(Header),
     BlockChunk(BlockChunk),
     BlockEnd(BlockEnd),
-    EndOfFile,
+    ExtractMetadata(ExtractMetadata),
+    ExtractChunk(ExtractChunk),
+    ExtractEnd(ExtractEnd),
+    EndOfFile(EndOfFile),
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -35,3 +38,26 @@ pub struct BlockEnd {
     pub crc32c: Option<u32>,
     pub xxh3: Option<u64>,
 }
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct ExtractMetadata {
+    pub has_content: bool,
+    pub file_path_components: Vec<String>,
+    pub is_truncated: bool,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct ExtractChunk {
+    #[serde(with = "serde_bytes")]
+    pub data: Vec<u8>,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct ExtractEnd {
+    pub crc32: Option<u32>,
+    pub crc32c: Option<u32>,
+    pub xxh3: Option<u64>,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct EndOfFile {}