+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+* Fixed: Verify functionality: block and payload digest checks were not functional.
+* Added: filter options for Extract command.
+* Added: extract option for Export command.
+* Changed: Made the EndOfFile message explicit for the Export and Import commands.
+
 ## 0.1.0 (2024-10-11)
 
-* First release
+* First release.